@@ -2,17 +2,6 @@
   "name": "graceful-process",
   "version": "1.2.0",
   "description": "graceful exit process even parent exit on SIGKILL.",
-  "files": [
-    "index.js",
-    "exit.js"
-  ],
-  "scripts": {
-    "lint": "eslint *.js test",
-    "test": "npm run lint && npm run test-local",
-    "test-local": "egg-bin test",
-    "cov": "cross-env COV=true egg-bin cov",
-    "ci": "npm run lint && npm run cov"
-  },
   "repository": {
     "type": "git",
     "url": "git@github.com:node-modules/graceful-process.git"
@@ -22,19 +11,59 @@
   },
   "author": "fengmk2",
   "license": "MIT",
-  "dependencies": {
-    "once": "^1.4.0",
-    "promise-timeout": "^1.3.0"
+  "engines": {
+    "node": ">= 18.19.0"
   },
+  "dependencies": {},
   "devDependencies": {
     "coffee": "^4.1.0",
     "cross-env": "^5.0.1",
-    "egg-bin": "^6.13.0",
-    "eslint": "^4.18.1",
-    "eslint-config-egg": "^7.0.0",
     "fkill": "^7.2.1",
     "mm": "^2.2.0",
     "mz-modules": "^2.1.0",
-    "urllib": "^2.26.0"
-  }
+    "urllib": "^2.26.0",
+    "@arethetypeswrong/cli": "^0.17.1",
+    "@eggjs/tsconfig": "1",
+    "@types/node": "22",
+    "@types/mocha": "10",
+    "egg-bin": "6",
+    "eslint": "8",
+    "eslint-config-egg": "14",
+    "tshy": "3",
+    "tshy-after": "1",
+    "typescript": "5"
+  },
+  "scripts": {
+    "lint": "eslint --cache src test --ext .ts",
+    "test": "npm run lint -- --fix && egg-bin test",
+    "ci": "npm run lint && egg-bin cov && npm run prepublishOnly && attw --pack",
+    "prepublishOnly": "tshy && tshy-after"
+  },
+  "type": "module",
+  "tshy": {
+    "exports": {
+      ".": "./src/index.ts",
+      "./package.json": "./package.json"
+    }
+  },
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
+      }
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "types": "./dist/commonjs/index.d.ts",
+  "main": "./dist/commonjs/index.js",
+  "module": "./dist/esm/index.js"
 }
